+++ src/Libraries/SmartStore.Services/Messages/IMessageTokenProvider.cs
@@ -16,9 +16,9 @@ namespace SmartStore.Services.Messages
     {
 		void AddStoreTokens(IList<Token> tokens, Store store);
 
-		void AddOrderTokens(IList<Token> tokens, Order order, int languageId);
+		void AddOrderTokens(IList<Token> tokens, Order order, Language language);
 
-        void AddShipmentTokens(IList<Token> tokens, Shipment shipment, int languageId);
+        void AddShipmentTokens(IList<Token> tokens, Shipment shipment, Language language);
 
         void AddOrderNoteTokens(IList<Token> tokens, OrderNote orderNote);
 
@@ -40,7 +40,7 @@ namespace SmartStore.Services.Messages
 
 		void AddProductTokens(IList<Token> tokens, Product product, Language language);
 
-		void AddForumTokens(IList<Token> tokens, Forum forum, int languageId);
+		void AddForumTokens(IList<Token> tokens, Forum forum, Language language);
 
         void AddForumTopicTokens(IList<Token> tokens, ForumTopic forumTopic,
             int? friendlyForumTopicPageIndex = null, int? appendedPostIdentifierAnchor = null);
@@ -55,13 +55,10 @@ namespace SmartStore.Services.Messages
 
         string[] GetListOfAllowedTokens();
 
-        //codehint: sm-add begin
         void AddBankConnectionTokens(IList<Token> tokens);
         
         void AddCompanyTokens(IList<Token> tokens);
 
         void AddContactDataTokens(IList<Token> tokens);
-        //codehint: sm-add end
-
     }
 }