+++ changelog.md
@@ -37,6 +37,7 @@
 * #800 Multi-store: Option to display all orders of all stores for customer in frontend
 * #457 Added option to hide the default image for categories and products
 * #451 Add message token for product shipping surcharge
+* #436 Make %Order.Product(s)% token to link the product detail page
 
 ### Improvements
 * (Perf) Implemented static caches for URL aliases and localized properties. Increases app startup and request speed by up to 30%.